# swagger.yaml
swagger: "2.0"
info:
  description: "TVJ Internal tool"
  version: "1.0.0"
  title: "TVJ Internal tool"
host: "localhost"
...
schemes:
- "https"
- "http"
paths:
  /user/login:
    post:
      tags:
      - "User"
      summary: "User Login"
      description: "User Login"
      operationId: "generateAuthenticationToken"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "body"
          name: "UserLoginReqDto"
          description: ""
          required: true
          schema:
            $ref: "#/definitions/UserLoginReqDto"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/UserLoginResDto"
        401:
          description: "Error code: ERR002"
          schema:
            $ref: "#/definitions/MessageResDto"
        423:
          description: "Error code: ERR007"
          schema:
            $ref: "#/definitions/MessageResDto"
  /user/save-last-logout:
    get:
      tags:
      - "User"
      summary: "Save user last logout"
      description: "Save user last logout"
      operationId: "Save user last logout"
      produces:
      - "application/json"
      responses:
        200:
          description: "Return code: SUC005"
          schema:
            $ref: "#/definitions/MessageResDto"
        400:
          description: "Return code: ERR011"
          schema:
            $ref: "#/definitions/MessageResDto" 
  /user/password-recover-send-request:
    post:
      tags:
      - "User"
      summary: "Recover password send request"
      description: "Recover password send request"
      operationId: "forgotPasswordSendRequest"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "body"
          name: "PasswordRecoverSendRequestReqDto"
          description: ""
          required: true
          schema:
            $ref: "#/definitions/PasswordRecoverSendRequestReqDto"
      responses:
        200:
          description: "Return code: SUC001"
          schema:
            $ref: "#/definitions/MessageResDto"
        408:
          description: "Return code: ERR002"
          schema:
            $ref: "#/definitions/MessageResDto"
  /user/password-recover-confirm-token:
    get:
      tags:
      - "User"
      summary: "Recover password confirm token"
      description: "Recover password confirm token"
      operationId: "forgotPasswordConfirmToken"
      produces:
      - "application/json"
      parameters:
      - name: "token"
        in: "query"
        description: "less than 100 characters"
        required: true
        type: "string"
      responses:
        200:
          description: "Return code: SUC004"
          schema:
            $ref: "#/definitions/MessageResDto"
        400:
          description: "Return code: ERR004"
          schema:
            $ref: "#/definitions/MessageResDto" 
  /user/password-recover-update-password:
    post:
      tags:
      - "User"
      summary: "Recover password update password"
      description: "Recover password update password"
      operationId: "recoverPasswordUpdatePassword"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "body"
          name: "PasswordRecoverUpdatePasswordReqDto"
          description: ""
          required: true
          schema:
            $ref: "#/definitions/PasswordRecoverUpdatePasswordReqDto"
      responses:
        200:
          description: "Return code: SUC002"
          schema:
            $ref: "#/definitions/MessageResDto"
        408:
          description: "Return code: ERR005"
          schema:
            $ref: "#/definitions/MessageResDto"
  /user/user-setting-get-info:
    get:
      tags:
      - "User"
      summary: "User setting get info"
      description: "User setting get info"
      operationId: "getUserSetting"
      produces:
      - "application/json"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/UserSettingResDto"
        400:
          description: "Return code: ERR009"
          schema:
            $ref: "#/definitions/MessageResDto" 
...
  /user/user-setting-update-info:
    put:
      tags:
      - "User"
      summary: "User setting update info"
      description: "User setting update info"
      operationId: "updateUserSetting"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "body"
          name: "UserSettingReqDto"
          description: ""
          required: true
          schema:
            $ref: "#/definitions/UserSettingUpdateReqDto"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/UserSettingUpdateReqDto"
        400:
          description: "Return code: ERR009"
          schema:
            $ref: "#/definitions/MessageResDto" 
  /user/user-setting-update-password:
    patch:
      tags:
      - "User"
      summary: "User setting update password"
      description: "User setting update password"
      operationId: "updatePassword"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "body"
          name: "UserSettingUpdatePasswordReqDto"
          description: ""
          required: true
          schema:
            $ref: "#/definitions/UserSettingUpdatePasswordReqDto"
      responses:
        200:
          description: "Return code: SUC002"
          schema:
            $ref: "#/definitions/MessageResDto"
        400:
          description: "Return code: ERR005"
          schema:
            $ref: "#/definitions/MessageResDto" 
  /user/upload-avatar:
    post:
      tags:
      - "User"
      summary: "User upload avatar"
      description: "User upload avatar"
      operationId: "uploadAvatar"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
        - in: "formData"
          name: file
          type: file
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/FileResDto"
        408:
          description: "Return code: ERR006"
          schema:
            $ref: "#/definitions/MessageResDto"
  /user/remove-avatar:
    delete:
      tags:
      - "User"
      summary: "User remove avatar"
      description: "User remove avatar"
      operationId: "removeAvatar"
      produces:
      - "application/json"
      responses:
        200:
          description: "Return code: SUC003"
          schema:
            $ref: "#/definitions/MessageResDto"
        408:
          description: "Return code: ERR010"
          schema:
            $ref: "#/definitions/MessageResDto"
...
definitions:
  UserLoginReqDto:
    type: "object"
    properties:
      username:
        type: string
        description: "Not Blank, less than 20 characters"
      password:
        type: string
        description: "Not Blank, between 8 and 20 characters"
  UserLoginResDto:
    type: "object"
    properties:
      token:
        type: string
        description: ""
      firstName:
        type: string
        description: ""
      lastName:
        type: string
        description: ""
      roleName:
        type: string
        description: ""
      isFirstTimeLogin:
        type: boolean
        description: ""
      avatar:
        type: string
        description: ""
  PasswordRecoverSendRequestReqDto:
    type: "object"
    properties:
      username:
        type: string
        description: "less than 20 characters" 
  UserSettingResDto:    
    type: "object"
    properties:
      username:
        type: string
        description: ""  
      firstName:
        type: string
        description: ""
      lastName:
        type: string
        description: ""
      email:
        type: string
        description: ""
      departmentId:
        type: string
        description: ""
      departmentName:
        type: boolean
        description: ""
      teamId:
        type: string
        description: ""
      teamName:
        type: string
        description: ""
      titleName:
        type: string
        description: ""
      address:
        type: string
        description: ""
      phone:
        type: boolean
        description: ""
      countryId:
        type: string
        description: ""
      countryName:
        type: string
        description: ""
      languageId:
        type: string
        description: ""
      languageName:
        type: string
        description: ""
      updatedDate:
        type: string
        description: ""
      statusId:
        type: string
        description: ""
      statusName:
        type: string
        description: ""
      avatar:
        type: string
        description: ""
  PasswordRecoverUpdatePasswordReqDto:
    type: "object"
    properties:
      token:
        type: string
        description: "Not Blank"
      newPassword:
        type: string
        description: "Not Blank, between 8 and 20 characters"
  UserSettingUpdateReqDto:
    type: "object"
    properties:
      teamId:
        type: string
        description: "between 1 and 40 characters"  
      address:
        type: string
        description: "less than 400 characters"
      phone:
        type: string
        description: "between 8 and 20 characters"
      countryId:
        type: string
        description: "between 1 and 40 characters"  
      languageId:
        type: boolean
        description: "Not Blank, between 1 and 40 characters"
      statusId:
        type: string
        description: "Not Blank, '1'-AVAILABLE, '2'-AWAY, '3'-BUSY, '4'-DO_NOT_DISTURB"
  UserSettingUpdatePasswordReqDto:
    type: "object"
    properties:
      oldPassword:
        type: string
        description: "Not Blank, between 8 and 20 characters"
      newPassword:
        type: string
        description: "Not Blank, between 8 and 20 characters"
  FileResDto:
    type: "object"
    properties:
      fileName:
        type: string
        description: ""    
...
  MessageResDto:
    type: "object"
    properties:
      code:
        type: string
        description: ""
      message:
        type: string
        description: ""